==> R/2.6_calculo-spatial-clustering.Rmd <==
---
title: "Cálculo de spatial clustering"
output: html_notebook
---

# Cargamos librerías 

```{r}
suppressMessages(library(tidyverse))
```

# Cargamos dataset con grifos y vecinos 

```{r}
grifos <- readRDS(here::here("data","processed","grifo_con_vecinos.rds"))

glimpse(grifos)
```

Tomemos un grifo cercano a mi casa y conocido para ver que esté bien la función anterior:

```{r}
grifos %>%
    filter(razon_social.princ == "INVERSIONES MAVU S.A.C.") %>%
    select(
        codigo_de_osinergmin.princ,
        razon_social.princ,
        bandera.princ,
        codigo_de_osinergmin.vec,
        razon_social.vec,
        bandera.vec
    )
```


# Funciones

Cargamos las funciones y hacemos algunas pruebas
```{r}

source(here::here("R","2.6_calculo-spatial-clustering_funciones.R"))


hallar_lista_vecinos(17947)
hallar_lista_vecinos(14665)
```


```{r}


hallar_cluster(8153, tipo = "razon_social")
hallar_cluster(14665)

```

```{r}


#cluster de 1
hallar_todos_cluster(14646, tipo = "razon_social")
#cluster de 2 
hallar_todos_cluster(19956)
#cluster independiente
hallar_todos_cluster(14665)



```

```{r}


#Prueba, debería retornar verdadero

comprobar_df_dup(hallar_todos_cluster(6743), hallar_todos_cluster(8332))

#debería retornar falso
comprobar_df_dup(hallar_todos_cluster(17944), hallar_todos_cluster(19956))

#nulo versus df
comprobar_df_dup(NULL, hallar_todos_cluster(19956))


```


Vemos que funciona, así que empaquetemos en una función:

```{r}



hallar_clusteres_mercado(19956)
hallar_clusteres_mercado(14645)

```

```{r}
cluster_14645_primax <- hallar_clusteres_mercado(14645)[[2]]
cluster_14645_primax
vec_cluster <- pull(cluster_14645_primax, codigo_de_osinergmin.princ)

grifos_coord <- readRDS(here::here("data","processed","grifo_coding_clean.rds")) %>% 
    filter(codigo_de_osinergmin %in% vec_cluster)
library(leaflet)

leaflet() %>%
    addMarkers(data = grifos_coord, 
               lat = ~ lat, 
               lng = ~ lon, 
               popup = paste0(grifos_coord$razon_social," ",grifos_coord$codigo_de_osinergmin),
               label = grifos_coord$bandera)  %>%
    addTiles()

```

Creamos función de spatial clustering (a nivel de banderas):

```{r}


calcular_sc(14645, "bandera")
calcular_sc(14645, "razon_social")
calcular_sc(19956)

hallar_clusteres_mercado(19956)
hallar_lista_vecinos(19956)
calcular_sc(19956, "razon_social")

```

# Test de la función principal: calcular_sc

La prueba la haremos con el grifo 14645 (Pecsa en el cruce de Cuba con Arenales). A la fecha de hoy (2019-03-31), se tiene para el grifo:
- N = 8
- M = 5 (Primax, dos de Pecsa, Repsol, Independiente):
    * `r k_repsol = 9`
    * `r k_primax_1 = 1`
    * `r k_primax_2 = 14`
    * `k_pecsa = 1`
    * `k_independiente = 1`
    
Por tanto el valor de sc es $sc = \frac{(9+1+14+1+1)/5}{8} = 0.65$. Verificamos que sea el mismo valor

```{r}
if (all.equal(calcular_sc(14645), 0.65)) {
    "Función calcular_sc pasa la prueba"
} else {
    stop("Función calcular_sc no pasa la prueba")

}
```
