==> R/2.7_creacion-database-spatial-clustering.Rmd <==
---
title: "Creación de database con spatial_clustering a nivel de razon social"
output: html_notebook
---

#Cargamos librerias necesarias y funciones

```{r}
suppressMessages(library(tidyverse))
source(here::here("R","2.6_calculo-spatial-clustering_funciones.R"))
library(purrr)

```

#Cargamos dataset con grifos y vecinos de Thiessen
Cargamos dos archivos generados previamente:
 - File con los vecinos de Thiessen generado en `r "2.5_estaciones-thiessen.Rmd"`. 
 - File con la lista de grifos y sus características generado en `r "1.6_base-datos-grifos-para-regresion.R"`

```{r}
grifos_coding_clean <- readRDS(here::here("data","processed","grifo_coding_clean.rds"))

```

# A nivel razón social

## Calculo pre-venta


```{r}
grifos <- readRDS(here::here("data","processed","grifo_con_vecinos_pre.RDS"))

grifos_sc <- grifos %>%
    distinct(codigo_de_osinergmin.princ) %>%
    mutate(sc = NA)

ptm <- proc.time()
sc_vector <- map_dbl(grifos_sc$codigo_de_osinergmin.princ, calcular_sc, "razon_social")
proc.time() - ptm

grifos_sc$sc <- sc_vector

grifos_sc_pre <- left_join(grifos_coding_clean, 
                           grifos_sc, 
                           c("codigo_de_osinergmin" = "codigo_de_osinergmin.princ")) 

```

# Calculo post-venta

Ahora, cambiamos las banderas que dicen Pecsa por Primax, 


```{r}
#cargamos la nueva base de datos con el nombre de grifos, de lo contrario no funcionan las funciones
grifos <- readRDS(here::here("data","processed","grifo_con_vecinos_post_razon.RDS")) 
#Cambiamos nombre de banderas
grifos_coding_clean_post <- grifos_coding_clean %>%
    mutate(razon_social = if_else(str_detect(razon_social, "PERUANA DE"),
                                  "COESTI S.A.", 
                                  razon_social))


```


```{r}
grifos_sc_post <- grifos %>%
    distinct(codigo_de_osinergmin.princ) %>%
    mutate(sc = NA)

ptm <- proc.time()
sc_vector_post <- map_dbl(grifos_sc_post$codigo_de_osinergmin.princ, calcular_sc, "razon_social")
proc.time() - ptm

grifos_sc_post$sc <- sc_vector_post

grifos_sc_post_1 <- left_join(grifos_coding_clean_post, 
                           grifos_sc_post, 
                           c("codigo_de_osinergmin" = "codigo_de_osinergmin.princ")) 

```
Unimos las medidas de sc pre y post en el mismo data file

```{r}
grifos_sc <- grifos_sc_pre %>%
    rename(sc_pre = sc)
grifos_sc$sc_post <- grifos_sc_post_1$sc

```


Grabamos el dataset para ser usado luego:

```{r}
saveRDS(grifos_sc, file = here::here("data","processed","grifos_con_sc_razon_social.RDS"))

```
