==> R/04_correccion-grifos-duplicados.Rmd <==
---
title: "Corrección grifos duplicados"
output: html_notebook
editor_options: 
  chunk_output_type: console
---
```{r}
library(tidyverse)
library(lubridate)
#Cargamos archivo de 02_data-wrangling.R
prices <- readRDS(here::here("data","processed","data_2005_2018_clean.rds"))
#Cargamos archivo de `04_importar-hoja-gdrive.R`
grifos <- readRDS(here::here("data","processed","grifo_coding_raw.rds"))
'%ni%' <- Negate('%in%')
```


# Corregimos los grifos duplicados, ayudandonos con la data de precios y la data geolocalizada.

Primero, todos los grifos que tienen mismo código, más de una entrada, y que no sean de un grifo conocido, esos los trataremos de manera independiente.

```{r}

#df con solo el primero de los repetidos (cualquiera)
grifos_repetidos <- grifos %>% 
    filter(!is.na(lat),
           !str_detect(razon_social,"COESTI"),
           !str_detect(razon_social,"REPSOL"),
           !str_detect(razon_social,"PERUANA DE COMB")) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1) %>%
    ungroup() %>%
    arrange(lat,codigo_de_osinergmin) %>%
    distinct(codigo_de_osinergmin, .keep_all = TRUE)

```

```{r}
#obtenmos un df sin ninguno de los repetidos
grifos_sin_marca_repetidos <- grifos %>% 
    filter(!is.na(lat),
           !str_detect(razon_social,"COESTI"),
           !str_detect(razon_social,"REPSOL"),
           !str_detect(razon_social,"PERUANA DE COMB")) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1) %>%
    ungroup()

#al df original, le quitamos todos los duplicados y le agregamos la lista depurada
grifos_1 <- anti_join(grifos,
                      grifos_sin_marca_repetidos,
                      by = NULL) %>%
    bind_rows(grifos_repetidos)

```


```{r}
#Verificamos
grifos_1 %>% 
    filter(!is.na(lat),
           !str_detect(razon_social,"COESTI"),
           !str_detect(razon_social,"REPSOL"),
           !str_detect(razon_social,"PERUANA DE COMB")) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1) %>%
    ungroup() %>%
    arrange(lat,codigo_de_osinergmin) %>%
    distinct(codigo_de_osinergmin, .keep_all = TRUE)
```


Ahora, `grifos_1` ya no tiene duplicados con el mismo código, al menos para casos sin grifos conocidos. Como hay grifos de PECSA que podrían haber pasado a COESTI, los filtramos y solo vemos duplicados que tengan que ver con Repsol. En caso duplicado, nos quedamos siempre con la razón social de Repsol:

```{r}
#nos quedamos solo con grifos repsol, cuando había otro con la misma razón social
#duplicado
grifos_repsol_depurado <- grifos_1 %>%
    filter(!is.na(lat),
           !str_detect(razon_social,"COESTI"),
           !str_detect(razon_social,"PERUANA DE COMB")) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n() > 1) %>%
    ungroup() %>%
    arrange(codigo_de_osinergmin) %>%
    filter(str_detect(razon_social, "REPSOL")) %>%
    distinct(codigo_de_osinergmin, .keep_all = TRUE)
```

```{r}
#Retiramos del df todos los 
grifos_repsol_repetidos <- grifos_1 %>%
    filter(!is.na(lat),
           !str_detect(razon_social,"COESTI"),
           !str_detect(razon_social,"PERUANA DE COMB")) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n() > 1) %>%
    ungroup() %>%
    arrange(codigo_de_osinergmin) 

grifos_2 <- anti_join(grifos_1,
                      grifos_repsol_repetidos,
                      by = NULL) %>%
    bind_rows(grifos_repsol_depurado)

#prueba
grifos_2 %>%
    filter(!is.na(lat),
           !str_detect(razon_social,"COESTI"),
           !str_detect(razon_social,"PERUANA DE COMB")) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n() > 1) %>%
    ungroup() %>%
    arrange(codigo_de_osinergmin) %>%
    filter(str_detect(razon_social, "REPSOL")) %>%
    distinct(codigo_de_osinergmin, .keep_all = TRUE)
```


Ahora debería haber menos grifos repetidos, verificamos:

```{r}
#Lo mismo para coesti
grifos_coesti_depurado <- grifos_2 %>% 
    filter(!is.na(lat)) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1) %>%
    ungroup() %>%
    arrange(codigo_de_osinergmin) %>%
    distinct(codigo_de_osinergmin, .keep_all = TRUE)

grifos_coesti_repetido <- grifos_2 %>% 
    filter(!is.na(lat)) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1) %>%
    ungroup() %>%
    arrange(codigo_de_osinergmin)

grifos_3 <- anti_join(grifos_2,
                      grifos_coesti_repetido,
                      by = NULL) %>%
    bind_rows(grifos_coesti_depurado)

grifos_3 %>% 
    filter(!is.na(lat)) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1) %>%
    ungroup() %>%
    arrange(codigo_de_osinergmin) %>%
    distinct(codigo_de_osinergmin, .keep_all = TRUE)
```
Ahora ya no hay duplicados por `codigo_de_osinergmin` en la base de datos.
```{r}
grifos_3 %>% 
    filter(!is.na(lat)) %>%
    group_by(codigo_de_osinergmin) %>%
    filter(n()>1)
```


## Verificamos por dirección duplicada

```{r}
grifos_3 %>%
    filter(!is.na(lat)) %>%
    group_by(direccion) %>%
    filter(n()>1)
```
OK, son grifos distintos, de un lado al otro de la avenida.

En adición, los grifos 83435 y 137688 son los mismos (compra de Repsol, dir AV. BAUZATE Y MEZA N° 1050). Nos quedamos con 137688

```{r}
grifos_3 %>%
    filter(codigo_de_osinergmin %in% c(83435,137688))

grifos_4 <- grifos_3 %>%
    filter(codigo_de_osinergmin != 83435)

prices_1 <- prices %>%
    mutate(codigo_de_osinergmin = as.numeric(codigo_de_osinergmin),
           codigo_de_osinergmin = if_else(codigo_de_osinergmin == 83435, 
                                          137688, 
                                          codigo_de_osinergmin))
```


## Verificamos que no hay grifos con las mismas coordenadas

```{r}
grifos_4 %>%
    filter(!is.na(lat)) %>%
    group_by(lat, lon) %>%
    filter(n()>1) %>%
    arrange(lat, lon) %>%
    ungroup() %>%
    select(codigo_de_osinergmin, razon_social, direccion, distrito, bandera)
```
Debemos tener cuidado con estos, y corregir la base de datos de precios, solo en este caso es necesario, (eliminaremos codigos de osinerming en grifos que son iguales.)

Los grifos 37795 y 132204 en realidad son los mismos (compra de Repsol). Por tanto, con quedamos con el código de Repsol y cambiamos en al base de datos de precios el código 37795 por 132204.


```{r}
prices_1 %>%
    filter(codigo_de_osinergmin == 37795)

prices_2 <- prices %>%
    mutate(codigo_de_osinergmin = as.numeric(codigo_de_osinergmin),
           codigo_de_osinergmin = if_else(codigo_de_osinergmin == 37795, 
                                          132204, 
                                          codigo_de_osinergmin))

prices_2 %>%
    filter(codigo_de_osinergmin == 37795)

#eliminamos de grifos:

grifos_5 <- grifos_4 %>%
    filter(codigo_de_osinergmin != 37795)

grifos_5 %>%
    filter(!is.na(lat)) %>%
    group_by(lat, lon) %>%
    filter(n()>1) %>%
    arrange(lat, lon) %>%
    ungroup() %>%
    select(codigo_de_osinergmin, razon_social, direccion, distrito, bandera)
```
Los grifos 9630 y 128675 en realidad son los mismos cambian de razón social. Nos quedamos con Delta Chocas (128675), por ser el nombre más reciente:

```{r}
prices_2 %>%
    filter(codigo_de_osinergmin == 9630)

prices_3 <- prices_2 %>%
    mutate(codigo_de_osinergmin = as.numeric(codigo_de_osinergmin),
           codigo_de_osinergmin = if_else(codigo_de_osinergmin == 9630, 
                                          128675, 
                                          codigo_de_osinergmin))

prices_3 %>%
    filter(codigo_de_osinergmin == 9630)

#eliminamos de grifos:

grifos_6 <- grifos_5 %>%
    filter(codigo_de_osinergmin != 9630)

grifos_6 %>%
    filter(!is.na(lat)) %>%
    group_by(lat, lon) %>%
    filter(n()>1) %>%
    arrange(lat, lon) %>%
    ungroup() %>%
    select(codigo_de_osinergmin, razon_social, direccion, distrito, bandera)
```


```{r}
prices %>%
    filter(codigo_de_osinergmin %in% c(9630,128675),
           year(fecha_hora) >= 2017) %>%
    ggplot2::ggplot(aes(x = fecha_hora, y = precio_de_venta, color = producto)) +
    ggplot2::geom_point() + 
    ggplot2::facet_wrap(~razon_social)
```
## Verificamos que no hay grifos evidentemente mal codificados con respecto a bandera

Verifiquemso que todos los grifos de Repsol estén bien codificados:

```{r}
grifos_6 %>% 
    filter(str_detect(razon_social,"REPSOL")) %>%
    select(codigo_de_osinergmin,razon_social, direccion, bandera) %>%
    mutate(bandera = str_to_upper(bandera)) %>%
    filter(bandera != "REPSOL")
```


Este es un caso super raro, en el que el grifo físicamente dice Pecsa pero está registrado a nombre de RECOSAC. Cambiaré el nombre de la bandera:

```{r}
grifos_7 <- grifos_6 %>%
    mutate(bandera = if_else(codigo_de_osinergmin == 31927, "REPSOL", bandera))

grifos_7 %>% 
    filter(str_detect(razon_social,"REPSOL")) %>%
    select(codigo_de_osinergmin,razon_social, direccion, bandera) %>%
    mutate(bandera = str_to_upper(bandera)) %>%
    filter(bandera != "REPSOL")
```

# Drop de grifos

Eliminamos el grifo 88228 EMPRESA DE TRANSPORTES EDILBERTO RAMOS S.A.C. por no tener observaciones:

```{r}
grifos_8 <- grifos_7 %>%
    filter(codigo_de_osinergmin != 88228)
```

# Verificamos las banderas

```{r}
grifos_8 %>% 
    count(razon_social, bandera, sort = TRUE) %>%
    filter(str_detect(razon_social, "REPSOL"))

grifos_9 <- grifos_8 %>% 
    mutate(bandera = case_when(
        str_detect(razon_social, "REPSOL") ~ "REPSOL",
        str_detect(razon_social, "COESTI") ~ "PRIMAX",
        str_detect(razon_social, "PERUANA DE ESTACIONES") ~ "PECSA",
        TRUE ~ bandera
    )) 
 
```


Guardamos el file de grifos y precios corregidos:

```{r}

saveRDS(grifos_9,
        file = here::here("data", "processed", "grifo_coding_raw_no_duplicate.rds"))

saveRDS(prices_3, 
        file = here::here("data", "processed", "data_prices_2005_2018_no_duplicates.rds"))

```
