==> R/06_base-datos-grifos-para-regresion.Rmd <==
---
title: "Base de datos de grifos limpia"
output: html_notebook
editor_options: 
  chunk_output_type: console
---

# Cargamos librerías y funciones

```{r}
suppressMessages(library(tidyverse))
library(lubridate)
library(stringr)
source(file = here::here("R", "funcion01_auxiliar-distancia.R"))
'%ni%' <- Negate('%in%')
```

Establecemos la distancia donde grifos compiten:

```{r}
distancia_competencia <- 1.5
```


# Cargamos los grifos

```{r remedy001}

grifos_datos_coding <- readRDS(here::here("data","processed","grifo_coding_raw_no_duplicate.rds")) %>%
    mutate(lat = as.double(lat))

glimpse(grifos_datos_coding)

grifos_datos_coding %>% 
    filter(str_detect(razon_social,"REPSOL")) %>%
    select(codigo_de_osinergmin,razon_social, direccion, bandera) %>%
    mutate(bandera = str_to_upper(bandera)) 
```


# Corregimos la columna de banderas

```{r remedy002}

grifos_datos_coding %>%
    mutate(bandera = str_to_upper(bandera, locale = "en")) %>%
    count(bandera, sort = TRUE) 

```

Recodificamos para que solo aparezcan las marcas reconocidas

```{r remedy003}
marcas <- c("REPSOL", "PRIMAX", "PETROPERU", "PECSA")
grifos_datos_coding_nombres <- grifos_datos_coding %>%
    mutate(bandera = str_to_upper(bandera),
           bandera = case_when(
               bandera %in% marcas ~ bandera,
               is.na(bandera) ~ NA_character_,
               TRUE ~ "INDEPENDIENTE"
           ),
           bandera = factor(bandera, levels = c("INDEPENDIENTE", "PETROPERU",
                                                "PECSA", "PRIMAX", "REPSOL")))

grifos_datos_coding_nombres %>%
    count(bandera)

```
Verificaciones:

```{r remedy004}

grifos_datos_coding_nombres %>% count(razon_social, bandera, sort = TRUE) %>%
    filter(str_detect(razon_social, "REPSOL"))

grifos_datos_coding_nombres %>% count(razon_social, bandera, sort = TRUE) %>%
    filter(str_detect(razon_social, "PERUANA DE ESTACIONES"))

grifos_datos_coding_nombres %>% count(razon_social, bandera, sort = TRUE) %>%
    filter(str_detect(razon_social, "COESTI"))

```
No hay errores aparentes en la codificación de la variable bandera.

# Añadimos variables

```{r}
grifo_datos_con_tipo <- grifos_datos_coding_nombres %>%
    mutate(
        tipo = case_when(
            is.na(bandera) ~ NA_character_,
            str_detect(razon_social, "REPSOL") ~ "PROPIA",
            str_detect(razon_social, "COESTI") ~ "PROPIA",
            str_detect(razon_social, "PERUANA DE ESTACIONES") ~ "PROPIA",
            bandera == "PETROPERU" ~ "ABANDERADA",
            bandera == "INDEPENDIENTE" ~ "INDEPENDIENTE",
            TRUE ~ "ABANDERADA"
        ),
        tipo = factor(tipo, levels = c("INDEPENDIENTE", "ABANDERADA", "PROPIA")),
        tipo_bandera = case_when(
            is.na(bandera) ~ NA_character_,
            str_detect(razon_social, "REPSOL") ~ "PROPIA REPSOL",
            str_detect(razon_social, "COESTI") ~ "PROPIA PRIMAX",
            str_detect(razon_social, "PERUANA DE ESTACIONES") ~ "PROPIA PECSA",
            bandera == "PETROPERU" ~ "ABANDERADA PETROPERU",
            bandera == "INDEPENDIENTE" ~ "INDEPENDIENTE",
            bandera == "REPSOL" ~ "ABANDERADA REPSOL",
            bandera == "PRIMAX" ~ "ABANDERADA PRIMAX",
            bandera == "PECSA" ~ "ABANDERADA PECSA"
        ),
        tipo_bandera = factor(tipo_bandera, 
                              levels = c("INDEPENDIENTE","ABANDERADA PETROPERU",
                                         "ABANDERADA PECSA", "ABANDERADA PRIMAX",
                                         "ABANDERADA REPSOL", "PROPIA PECSA",
                                         "PROPIA PRIMAX", "PROPIA REPSOL")),
        con_glp = case_when(
            is.na(glp) ~ NA_real_,
            glp >= 1 ~ 1,
            TRUE ~ 0),
        con_glp = as.factor(con_glp),
        con_gnv = case_when(
            is.na(gnv) ~ NA_real_,
            gnv >= 1 ~ 1,
            TRUE ~ 0),        
        con_gnv = as.factor(con_gnv),
    ) %>%
    select(-codificador, -ruc, -n, -revisar)
    

glimpse(grifo_datos_con_tipo)

```

# Vemos los grifos más cercanos 

```{r remedy005}

grifos_con_datos <- grifo_datos_con_tipo %>%
    filter(!is.na(lat))

grifos_con_datos

```

Hallamos la distancia al grifo más cercano

```{r remedy006}
matriz_distancias <- grifos_con_datos %>%
    select("name" = codigo_de_osinergmin , lat, lon) %>%
    distinct() %>%
    GeoDistanceInMetresMatrix(.) / 1000

```

Generamos un archivo con las distancias para revisar manualmente errores de codificacion

```{r remedy007}

df_distancias <- as_tibble(matriz_distancias, rownames = "codigo_de_osinergmin") %>%
    gather(key = "grifo_distancia", value = "distancia", -codigo_de_osinergmin) %>%
    filter(distancia != 0) %>%
    group_by(codigo_de_osinergmin) %>%
    mutate(distancia_min = min(distancia)) %>%
    ungroup() %>%
    mutate(codigo_de_osinergmin = as.numeric(codigo_de_osinergmin)) %>%
    arrange(as.numeric(codigo_de_osinergmin)) %>%
    filter(distancia == distancia_min) %>%
    arrange(distancia_min) 

left_join(
    df_distancias,
    select(
        grifos_con_datos,
        codigo_de_osinergmin,
        razon_social,
        distrito,
        direccion,
        bandera,
        lat, lon
    ),
    by = "codigo_de_osinergmin"
) %>%
    unite(coordenadas, lat, lon, sep = ", ") %>% 
    write_excel_csv(file = here::here("data", "processed", "temporal", "grifos-distancia.csv"))

```



# Creamos variable con distancia mínima al grifo competidor más cercano

```{r remedy008}
#' Pendiente, como hacer para que no considere si el grifo más cercano tiene el mismo ruc
grifos_distancia_minima <- as_tibble(matriz_distancias, rownames = "codigo_de_osinergmin") %>%
    gather(key = "grifo_distancia", value = "distancia", -codigo_de_osinergmin) %>%
    filter(distancia != 0) %>%
    group_by(codigo_de_osinergmin) %>%
    mutate(distancia_min = min(distancia)) %>%
    ungroup() %>%
    arrange(as.numeric(codigo_de_osinergmin)) %>%
    filter(distancia == distancia_min) %>%
    select(-distancia) %>% 
    mutate_at(c("codigo_de_osinergmin", "grifo_distancia"), as.numeric)

```

# Distancia promedio a grifos a menos de `distancia_competencia`

Creamos df con dist promedio a grifos, número de grifos vecinos (1.5 km), y si el grifo tiene grifo propio de Pecsa vecino.

```{r}
grifos_distancia_pecsa <- as_tibble(matriz_distancias, rownames = "codigo_de_osinergmin") %>%
  gather(key = "grifo_distancia", value = "distancia", -codigo_de_osinergmin) %>%
  filter(distancia != 0, distancia < distancia_competencia) %>%
  mutate(
    codigo_de_osinergmin = as.numeric(codigo_de_osinergmin),
    grifo_distancia = as.numeric(grifo_distancia)
  ) %>%
  arrange(codigo_de_osinergmin) %>%
  left_join(
    grifos_con_datos %>% select(codigo_de_osinergmin, tipo_bandera),
    by = c("grifo_distancia" = "codigo_de_osinergmin")
  ) %>%
  group_by(codigo_de_osinergmin) %>%
  mutate(
    distancia_avg = mean(distancia),
    num_grifos_cerc = n(),
    vecino_pecsa_dist = if_else(tipo_bandera == "PROPIA PECSA",
      1,
      0
    ),
    num_vecinos_pecsa = sum(vecino_pecsa_dist)
  ) %>%
  arrange(codigo_de_osinergmin, desc(vecino_pecsa_dist)) %>%
  distinct(codigo_de_osinergmin,
    distancia_avg,
    num_grifos_cerc,
    .keep_all = TRUE
  ) %>%
  select(-distancia, -tipo_bandera, -grifo_distancia)
  
grifos_distancia_primax <- as_tibble(matriz_distancias, rownames = "codigo_de_osinergmin") %>%
  gather(key = "grifo_distancia", value = "distancia", -codigo_de_osinergmin) %>%
  filter(distancia != 0, distancia < distancia_competencia) %>%
  mutate(
    codigo_de_osinergmin = as.numeric(codigo_de_osinergmin),
    grifo_distancia = as.numeric(grifo_distancia)
  ) %>%
  arrange(codigo_de_osinergmin) %>%
  left_join(
    grifos_con_datos %>% select(codigo_de_osinergmin, tipo_bandera),
    by = c("grifo_distancia" = "codigo_de_osinergmin")
  ) %>%
  group_by(codigo_de_osinergmin) %>%
  mutate(
    vecino_primax_dist = if_else(tipo_bandera == "PROPIA PRIMAX",
      1,
      0
    ),
    num_vecinos_primax = sum(vecino_primax_dist)
  ) %>%
  arrange(codigo_de_osinergmin, desc(vecino_primax_dist)) %>%
  distinct(codigo_de_osinergmin,
    .keep_all = TRUE
  ) %>%
  select(-distancia, -tipo_bandera, -grifo_distancia)    
```

Lo agregamos al archivo anterior

```{r remedy010}
grifos_distancias <- full_join(grifos_distancia_minima, grifos_distancia_pecsa,
                               by = "codigo_de_osinergmin") %>%
  left_join(grifos_distancia_primax, by = "codigo_de_osinergmin") %>% 
    rename("grifo_mas_cercano" = grifo_distancia) 

grifo_full <-
    full_join(grifos_con_datos,
              grifos_distancias,
              by = c("codigo_de_osinergmin"))
#los grifos que no tienen competencia en 1.5 km hacemos lo siguiente:

grifo_full <- grifo_full %>% 
    mutate(num_grifos_cerc = if_else(is.na(num_grifos_cerc), 0L, num_grifos_cerc),
           distancia_avg = if_else(is.na(distancia_avg), 1.5, distancia_avg),
           num_vecinos_pecsa = as.integer(num_vecinos_pecsa),
           num_vecinos_pecsa = if_else(is.na(num_vecinos_pecsa), 0L, num_vecinos_pecsa),
           vecino_pecsa_dist = if_else(is.na(vecino_pecsa_dist), 0, vecino_pecsa_dist),
            num_vecinos_primax = as.integer(num_vecinos_primax),
           num_vecinos_primax = if_else(is.na(num_vecinos_primax), 0L, num_vecinos_primax),
           vecino_primax_dist = if_else(is.na(vecino_primax_dist), 0, vecino_pecsa_dist)
           )
```
# Corregimos razones sociales de marcas principales:

```{r}
grifo_full_1 <- grifo_full %>%
    mutate(
        razon_social = case_when(
            str_detect(razon_social, "COESTI") ~ "COESTI S.A.",
            str_detect(razon_social, "REPSOL") ~ "REPSOL COMERCIAL S.A.C.",
            TRUE ~ razon_social
        )
    )
        
```

# Arreglamos tipos de variables

Primero vemos que tipos de variables tenemos

```{r}
skimr::skim(grifo_full_1)
```

```{r}
grifo_full_2 <- grifo_full_1 %>% 
  mutate(
    tiene_mecanico = if_else(mecanico + aceite + lavado > 0, 1, 0),
    con_glp = as.numeric(con_glp) - 1,
    con_gnv = as.numeric(con_gnv) - 1
  )

grifo_full_2 %>% 
  filter(distrito == "SAN JUAN DE LURIGANCHO")
```

# Realizamos la matriz de distancias a los puntos de venta

```{r}
lista_terminales <- c("Callao","Pampilla","Conchan")
terminales <- tibble(name = lista_terminales, 
       lat = c(-12.038859641394353, -11.919635654091762, -12.2538367359577),
       lon = c(-77.12722709481619, -77.12977045738286, -76.92312446782478) )

matriz_distancias_terminal <- grifo_full_2 %>%
  select("name" = codigo_de_osinergmin , lat, lon) %>%
  distinct() %>%
  mutate(name = as.character(name)) %>% 
  bind_rows(terminales) %>% 
  GeoDistanceInMetresMatrix(.) / 1000

df_distancias_terminal <- as_tibble(matriz_distancias_terminal, 
                           rownames = "codigo_de_osinergmin") %>%
  gather(key = "grifo_distancia", value = "distancia", -codigo_de_osinergmin) %>%
  filter(distancia != 0,
         grifo_distancia %in% lista_terminales,
         codigo_de_osinergmin %ni% lista_terminales) %>% 
  arrange(codigo_de_osinergmin) %>% 
  group_by(codigo_de_osinergmin) %>%
  mutate(distancia_min_term = min(distancia)) %>%
  mutate(distancia_prom_term = mean(distancia)) %>% 
  ungroup() %>%
  mutate(codigo_de_osinergmin = as.numeric(codigo_de_osinergmin)) %>%
  arrange(as.numeric(codigo_de_osinergmin)) %>%
  filter(distancia == distancia_min_term) %>%
  arrange(distancia_min_term) %>% 
  select(-distancia) %>% 
  rename(terminal_cercano = grifo_distancia)


grifo_full_3 <- grifo_full_2 %>% 
  left_join(df_distancias_terminal, by = "codigo_de_osinergmin")


```


#Guardamos
```{r remedy011}

saveRDS(grifo_full_3, file = here::here("data","processed","grifo_coding_clean.rds"))


```
