==> R/07_calculo_vecinos_thiessen.Rmd <==
---
title: "Polígonos Thiessen para dataset"
output: 
  html_notebook: 
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r}
library(sf)
library(dismo)
suppressMessages(library(tidyverse))
suppressMessages(library(spatstat))
suppressMessages(library(deldir))
library(sp)
library(leaflet)
source(here::here("R","funcion02_thiessen.R"))
conflicted::conflict_prefer("select", "dplyr")
conflicted::conflict_prefer("filter", "dplyr")
```


Se construirán los polígonos de Thiesen para todo el dataset, así como se identificarán los vecinos y de que dueño y bandera son. Además, se construirán gráficos de los vecinos

Cargamos la base de datos con grifos y ubicaciones:

```{r}
grifos <- readRDS(here::here("data","processed","grifo_coding_clean.rds"))
glimpse(grifos)
```


# Mapa en google maps - Visualizaci?n

Usamos los siguientes recursos:
- https://stackoverflow.com/questions/43055894/how-to-assign-the-data-of-a-centroid-marker-to-the-voronoi-thiessen-polygon-it
- https://rpubs.com/walkerke/rstudio_gis

Replicando para mi data con el primer link:

```{r}
grifos_puntos <- grifos %>%
  st_as_sf(crs = 4326, coords = c("lon", "lat"))

plot(grifos_puntos["bandera"], key.pos = 1, axes = FALSE,key.width = lcm(1.3), key.length = 1.0)
```

Siguiente:

```{r}
polys <- grifos_puntos %>%
  as("Spatial") %>%
  voronoi() %>%
  st_as_sf() %>%
  st_set_crs(., 4326)

plot(polys["bandera"],key.pos = 1, axes = FALSE,key.width = lcm(1.3), key.length = 1.0)
```


Now, visualize it with Leaflet using your desired colors:

```{r}
pal <- colorFactor("plasma", polys$bandera)


polys %>%
  leaflet() %>%
  addProviderTiles(providers$CartoDB.Positron) %>%
  addPolygons(fillColor = ~pal(bandera), weight = 0.5, color = "grey") %>%
  addCircleMarkers(data = grifos_puntos, label = ~bandera, color = ~pal(bandera))
```
Todo bien, pero al coger un rectángulo toma área del mar y no tiene sentido. Intentaré arreglar eso




# Corrigiendo los límites (para que no sea una caja y no tome el mar)

Importamos límites de la costa para ?rea de inter?s (los saqu? manualmente):

```{r}
coordenadas_costa <-
    readr::read_csv(here::here("data", "limites_costa.csv"),
                    col_types = c("cc")) %>%
    separate(
        `lat, long`,
        into = c("y", "x"),
        sep = ",",
        convert = TRUE
    )


```

Creamos un gr?fico simple para ver si funciona (ojo primero va la longitud, luego latitud)

```{r}

W <- owin(poly=data.frame(x=(coordenadas_costa$x), y=(coordenadas_costa$y)))
plot(W)

grifos_coord <- grifos %>% 
    dplyr::select(x = lon, y = lat)

X <- as.ppp(grifos_coord,W=W) 
plot(X) # Just to make sure it looks right. 
dX <- dirichlet(X) 
plot(dX) # Just to make sure ..... 
```
Sacamos una gráfica para la tesis

```{r fig.height=10, fig.width=10}
# grifos_coord <- grifos %>%
#   dplyr::filter(distrito == "SAN ISIDRO") %>%
#   dplyr::select(x = lon, y = lat)
# 
# W <- owin(xrange = c(min(grifos_coord$x)-0.001, max(grifos_coord$x)+0.001),
#           yrange = c(min(grifos_coord$y)-0.01, max(grifos_coord$y)+0.01))
# 
# 
# X <- as.ppp(grifos_coord,W)
# plot(X) # Just to make sure it looks right.
# dX <- dirichlet(X)
# par(new=TRUE)
# pdf("rplot.pdf") 
# plot(dX) # Just to make sure .....
# dev.off() 
```



Ahora lo hacemos en mapa interactivo usando leaflet
```{r}
library(maptools)
dXsp <- as(dX, "SpatialPolygons")

dsp_df <- SpatialPolygonsDataFrame(dXsp, 
                                   data = data.frame(id = 1:length(dXsp@polygons)))

proj4string(dsp_df) <- CRS("+proj=longlat +datum=WGS84")
dsp_xy <- spTransform(dsp_df, CRS("+proj=longlat +datum=WGS84"))


#Nos quedamos solo con los distintos, y agregamos columna de datos de bandera:

grifos_filtrado <- grifos %>% 
  distinct(lat,lon, .keep_all = TRUE) %>%
  mutate(tipo_bandera = recode_factor(tipo_bandera, 
                                      "ABANDERADA PECSA" = "OTRA",
                                      "ABANDERADA PRIMAX" = "OTRA",
                                      "ABANDERADA REPSOL" = "OTRA",
                                      "INDEPENDIENTE" = "OTRA",
                                      "ABANDERADA PETROPERU" = "OTRA",
                                      "PROPIA PECSA" = "PERUANA DE COMBUSTIBLES (PECSA)",
                                      "PROPIA REPSOL" = "REPSOL COMERCIAL",
                                      "PROPIA PRIMAX" = "COESTI (PRIMAX)"
                                      )
         ) %>% 
  rownames_to_column("rowname")

grifos_thissen <- merge(dsp_xy, grifos_filtrado, by.x="id", by.y="rowname")
saveRDS(grifos_thissen, here::here("data", "processed", "shape_file_grifos_thiessen.rds"))
```

Gráfico usando leaflet

```{r}


pal <- colorFactor("plasma", unique(grifos_thissen$bandera))
pal2 <- colorFactor(c("white", "green", "red", "orange", "blue"), unique(grifos_thissen$bandera))
pal3 <- colorFactor(c("white", "green", 
                      "darksalmon", "lightgoldenrodyellow", "dodgerblue",
                      "red", "orange", "blue"), unique(grifos_thissen$tipo_bandera))
pal4 <- colorFactor("plasma", unique(grifos_thissen$tipo_bandera)) 
leaflet() %>%
  addMarkers(data = grifos_filtrado, 
             lat = ~ lat, 
             lng = ~ lon, 
             popup = paste0(grifos$razon_social," ",grifos$codigo_de_osinergmin),
             label = grifos_thissen$tipo_bandera,
             labelOptions = labelOptions(noHide = F))  %>%
  # addPolygons(data = dsp_xy, 
  #             color = "green", 
  #             fill = "green") %>%
  addPolygons(data = grifos_thissen,
              #fillColor = ~pal2(bandera), 
              fillColor = ~pal4(tipo_bandera),
              weight = 1, 
              color = "black") %>%
  addTiles()
```



# Pruebas

```{r}
grifos_thissen$codigo_de_osinergmin
grifos_thissen[grifos_thissen$codigo_de_osinergmin == 15234,]

for (grifo in grifos_thissen$codigo_de_osinergmin) {
  raster::intersect(grifos_thissen[grifos_thissen$codigo_de_osinergmin == 15234,], 
                    grifos_thissen[grifos_thissen$codigo_de_osinergmin == grifo,])
}
raster::intersect(grifos_thissen[grifos_thissen$codigo_de_osinergmin == 7709,], 
                  grifos_thissen[grifos_thissen$codigo_de_osinergmin == 40975,])

rgeos::gIntersection(grifos_thissen[grifos_thissen$codigo_de_osinergmin == 7709,], 
                  grifos_thissen[grifos_thissen$codigo_de_osinergmin == 40975,])

rgeos::gIntersection(grifos_thissen[grifos_thissen$codigo_de_osinergmin == 7709,], 
                  grifos_thissen[grifos_thissen$codigo_de_osinergmin == 9536,])

vec_15234 <- vector()
for (grifo in grifos_thissen$codigo_de_osinergmin) {
  
  t <- rgeos::gIntersection(grifos_thissen[grifos_thissen$codigo_de_osinergmin == 15234,], 
                       grifos_thissen[grifos_thissen$codigo_de_osinergmin == grifo,])
  
  if (!is.null(t)) {
    vec_15234 <- append(vec_15234, grifo)
  }
}
```




# Para antes del merger

```{r}

hallar_bandera <- function(cod1) {
    # se debe tener cargado el ouput de 2.5 estaciones thiessen
    #input: codigo del grifo
    #ouput: bandera del grifo
  unique(grifos[grifos$codigo_de_osinergmin == cod1, ]$bandera)
    # grifos %>% 
    #     filter(codigo_de_osinergmin == cod1) %>%
    #     distinct(bandera) %>% 
    #     pull()
}

vec_thiessen_irr <- function(cod) {
  #función que devuelve un vector con los vecinos del grifo
  ans <- vector()
  for (grifo in grifos_thissen$codigo_de_osinergmin) {
    
    t <- rgeos::gIntersection(grifos_thissen[grifos_thissen$codigo_de_osinergmin == cod,], 
                              grifos_thissen[grifos_thissen$codigo_de_osinergmin == grifo,])
    
    if (!is.null(t)) {
      ans <- append(ans, grifo)
    }
  }
  bandera <- hallar_bandera(cod)
  tibble("codigo_de_osinergmin.princ" = rep(cod, length(ans)), 
       "codigo_de_osinergmin.vec" = ans) # %>%
    # distinct() %>%
    # mutate(bandera.princ = bandera,
    #        bandera.vec = hallar_bandera(codigo_de_osinergmin.vec))
}


```


```{r}


vec_thiessen_irr(19963)
```

Ahora lo hacemos para toda la muestra:

```{r}
grifos_thiessen_irr <- tibble("codigo_de_osinergmin.princ" =  numeric(), 
       "codigo_de_osinergmin.vec" = numeric())
grifos_thiessen_irr
# for (grifo in grifos$codigo_de_osinergmin) {
#   
#   grifos_thiessen_irr <- bind_rows(grifos_thiessen_irr,
#                                    vec_thiessen_irr(grifo) )  
#   
#   }

list_thiessen <- purrr::map_dfr(.x = grifos$codigo_de_osinergmin, .f = vec_thiessen_irr)

grifos_thiessen_irr <- list_thiessen %>%
  filter(codigo_de_osinergmin.princ != codigo_de_osinergmin.vec)

saveRDS(grifos_thiessen_irr, here::here("data", "processed","grifos_thiessen_irr.Rds"))
```

Ahora agregamos información sobre las banderas:

```{r}
grifos_thiessen_irr <- readRDS(here::here("data", "processed","grifos_thiessen_irr.Rds"))

grifos_con_vecinos <- left_join(
  grifos_thiessen_irr,
  dplyr::select(
    grifos,
    codigo_de_osinergmin,
    bandera.princ = bandera,
    razon_social.princ = razon_social,
    tipo_bandera.princ = tipo_bandera
  ),
  by = c("codigo_de_osinergmin.princ" = "codigo_de_osinergmin")
) %>%
  left_join(
    dplyr::select(
      grifos,
      codigo_de_osinergmin,
      bandera.vec = bandera,
      razon_social.vec = razon_social,
      tipo_bandera.vec = tipo_bandera
    ),
    by = c("codigo_de_osinergmin.vec" = "codigo_de_osinergmin")
  )
```

```{r}
grifos_con_vecinos %>%
  filter(razon_social.princ == "INVERSIONES MAVU S.A.C.") %>%
  select(ends_with("vec"))


```


```{r}
saveRDS(grifos_con_vecinos, file = here::here("data","processed","grifo_con_vecinos_pre.RDS"))
```


#Cargamos dataset con grifos y vecinos de Thiessen
Cargamos dos archivos generados previamente:
 - File con los vecinos de Thiessen generado en `r "2.5_estaciones-thiessen.Rmd"`. 
 - File con la lista de grifos y sus características generado en `r "1.6_base-datos-grifos-para-regresion.R"`

```{r}
grifos_coding_clean <- readRDS(here::here("data","processed","grifo_coding_clean.rds"))

```


# Calculamos los vecinos de thiessen

```{r}
grifos_vecinos <- readRDS(here::here("data", "processed", "grifo_con_vecinos_pre.RDS"))

vecinos_pecsa_thissen <- grifos_vecinos %>%
    group_by(codigo_de_osinergmin.princ) %>%
    mutate(vecino_pecsa_thiessen = if_else(str_detect(razon_social.vec, "PERUANA DE ESTACIONES"),
                                           1,
                                           0
    )) %>%
    arrange(codigo_de_osinergmin.princ, desc(vecino_pecsa_thiessen)) %>%
    distinct(codigo_de_osinergmin.princ, .keep_all = TRUE) %>%
    select("codigo_de_osinergmin" = codigo_de_osinergmin.princ, vecino_pecsa_thiessen)

vecinos_primax_thissen <- grifos_vecinos %>%
    group_by(codigo_de_osinergmin.princ) %>%
    mutate(vecino_primax_thiessen = if_else(str_detect(razon_social.vec, "COESTI"),
                                           1,
                                           0
    )) %>%
    arrange(codigo_de_osinergmin.princ, desc(vecino_primax_thiessen)) %>%
    distinct(codigo_de_osinergmin.princ, .keep_all = TRUE) %>%
    select("codigo_de_osinergmin" = codigo_de_osinergmin.princ, vecino_primax_thiessen)

market_shares <- grifos_vecinos %>%
  group_by(codigo_de_osinergmin.princ) %>%
  mutate(
    vecino_pecsa_thiessen = if_else(str_detect(razon_social.vec, "PERUANA DE ESTACIONES"),
      1,
      0
    ),
    num_vecinos_thiessen = n(),
    num_vecinos_independientes = sum(bandera.vec == "INDEPENDIENTE"),
    num_vecinos_pecsa = sum(tipo_bandera.vec == "PROPIA PECSA"),
    num_vecinos_primax = sum(tipo_bandera.vec == "PROPIA PRIMAX"),
    num_vecinos_repsol = sum(tipo_bandera.vec == "PROPIA REPSOL"),
    market_share_antes = case_when(
      bandera.princ == "INDEPENDIENTE" ~ 1 / (num_vecinos_thiessen + 1),
      tipo_bandera.princ == "PROPIA PECSA" ~ (1 + num_vecinos_pecsa) / (num_vecinos_thiessen + 1),
      tipo_bandera.princ == "PROPIA PRIMAX" ~ (1 + num_vecinos_primax) / (num_vecinos_thiessen + 1),
      tipo_bandera.princ == "PROPIA REPSOL" ~ (1 + num_vecinos_repsol) / (num_vecinos_thiessen + 1),
      TRUE ~ 1 / (num_vecinos_thiessen + 1)
    ),
    market_share_despues = case_when(
      bandera.princ == "INDEPENDIENTE" ~ 1 / (num_vecinos_thiessen + 1),
      tipo_bandera.princ == "PROPIA PECSA" ~ (1 + num_vecinos_pecsa + num_vecinos_primax) / (num_vecinos_thiessen + 1),
      tipo_bandera.princ == "PROPIA PRIMAX" ~ (1 + num_vecinos_primax + num_vecinos_pecsa) / (num_vecinos_thiessen + 1),
      tipo_bandera.princ == "PROPIA REPSOL" ~ (1 + num_vecinos_repsol) / (num_vecinos_thiessen + 1),
      TRUE ~ 1 / (num_vecinos_thiessen + 1)
    )
  ) %>% 
  select(codigo_de_osinergmin.princ, market_share_antes, market_share_despues, starts_with("num_vecinos")) %>% 
  distinct()

grifos_market_share <- grifos_vecinos  %>% 
  left_join(market_shares, by = c("codigo_de_osinergmin.vec" = "codigo_de_osinergmin.princ")) %>% 
  select(-starts_with("bandera"), -starts_with("razon_social")) %>% 
  group_by(codigo_de_osinergmin.princ) %>% 
  mutate(HHI_vecinos_antes = sum((market_share_antes*100)^2),
         HHI_vecinos_despues = sum((market_share_despues*100)^2)) %>% 
  ungroup() %>% 
  select(-starts_with("market_share"), -ends_with(".vec")) %>% 
  left_join(market_shares, by = c("codigo_de_osinergmin.princ" = "codigo_de_osinergmin.princ")) %>% 
  distinct(codigo_de_osinergmin.princ, .keep_all = T) %>% 
  group_by(codigo_de_osinergmin.princ) %>% 
  mutate(HHI_antes = HHI_vecinos_antes + (100*market_share_antes)^2,
         HHI_despues = HHI_vecinos_despues + (100*market_share_despues)^2) %>% 
  select(-starts_with("HHI_vecinos"), codigo_de_osinergmin = codigo_de_osinergmin.princ,
         -tipo_bandera.princ) 
    
vecinos_primax_thissen <- grifos_vecinos %>%
    group_by(codigo_de_osinergmin.princ) %>%
    mutate(vecino_primax_thiessen = if_else(str_detect(razon_social.vec, "COESTI"),
                                           1,
                                           0
    )) %>%
    arrange(codigo_de_osinergmin.princ, desc(vecino_primax_thiessen)) %>%
    distinct(codigo_de_osinergmin.princ, .keep_all = TRUE) %>%
    select("codigo_de_osinergmin" = codigo_de_osinergmin.princ, vecino_primax_thiessen)

grifos_data_final <- grifos_coding_clean %>% 
  left_join(vecinos_pecsa_thissen, by = "codigo_de_osinergmin") %>% 
  left_join(vecinos_primax_thissen, by = "codigo_de_osinergmin") %>% 
  left_join(grifos_market_share, by = "codigo_de_osinergmin")

saveRDS(grifos_data_final, file = here::here("data","processed","grifos_data_final.RDS"))

```

# A nivel razón social

## Calculo pre-venta


```{r}
grifos <- readRDS(here::here("data","processed","grifo_con_vecinos_pre.RDS"))

grifos_sc <- grifos %>%
    distinct(codigo_de_osinergmin.princ) %>%
    mutate(sc = NA)
# 
# ptm <- proc.time()
# sc_vector <- map_dbl(grifos_sc$codigo_de_osinergmin.princ, calcular_sc, "razon_social")
# proc.time() - ptm
# 
# grifos_sc$sc <- sc_vector

grifos_pre <- left_join(grifos_coding_clean, 
                           grifos_sc, 
                           c("codigo_de_osinergmin" = "codigo_de_osinergmin.princ")) 

```


Grabamos el dataset para ser usado luego:

```{r}
saveRDS(grifos_pre, file = here::here("data","processed","grifos_con_sc_razon_social.RDS"))

```

<!-- # Ahora post merger a nivel banderas -->
<!-- ```{r} -->
<!-- grifos_post <- grifos %>% -->
<!--   mutate(bandera = recode_factor(bandera, "PECSA" = "PRIMAX")) -->
<!-- ``` -->

<!-- ```{r} -->


<!-- grifos_con_vecinos_post <- left_join( -->
<!--   grifos_thiessen_irr, -->
<!--   select( -->
<!--     grifos_post, -->
<!--     codigo_de_osinergmin, -->
<!--     bandera.princ = bandera, -->
<!--     razon_social.princ = razon_social -->
<!--   ), -->
<!--   by = c("codigo_de_osinergmin.princ" = "codigo_de_osinergmin") -->
<!-- ) %>% -->
<!--   left_join( -->
<!--     select( -->
<!--       grifos_post, -->
<!--       codigo_de_osinergmin, -->
<!--       bandera.vec = bandera, -->
<!--       razon_social.vec = razon_social -->
<!--     ), -->
<!--     by = c("codigo_de_osinergmin.vec" = "codigo_de_osinergmin") -->
<!--   ) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- saveRDS(grifos_con_vecinos_post, file = here::here("data","processed","grifo_con_vecinos_post_bandera.RDS")) -->
<!-- ``` -->

<!-- # Ahora post merger a nivel razon_social -->

<!-- Cambiaré la razón social de los grifos Pecsa a Coesti, para efectos del cálculo (se manejan como empresas distintas) -->

<!-- ```{r} -->
<!-- grifos_thiessen_irr <- readRDS(here::here("data", "processed","grifos_thiessen_irr.Rds")) -->

<!-- library(stringr) -->
<!-- grifos_post_razon <- grifos %>% -->
<!--   mutate(razon_social = if_else(str_detect(razon_social, "PERUANA DE"), -->
<!--                                 "COESTI S.A.", -->
<!--                                 razon_social)) -->

<!-- grifos_con_vecinos_post_razon <- left_join( -->
<!--   grifos_thiessen_irr, -->
<!--   select( -->
<!--     grifos_post_razon, -->
<!--     codigo_de_osinergmin, -->
<!--     bandera.princ = bandera, -->
<!--     razon_social.princ = razon_social -->
<!--   ), -->
<!--   by = c("codigo_de_osinergmin.princ" = "codigo_de_osinergmin") -->
<!-- ) %>% -->
<!--   left_join( -->
<!--     select( -->
<!--       grifos_post_razon, -->
<!--       codigo_de_osinergmin, -->
<!--       bandera.vec = bandera, -->
<!--       razon_social.vec = razon_social -->
<!--     ), -->
<!--     by = c("codigo_de_osinergmin.vec" = "codigo_de_osinergmin") -->
<!--   ) -->
<!-- ``` -->




<!-- ```{r} -->
<!-- saveRDS(grifos_con_vecinos_post_razon,  -->
<!--         file = here::here("data","processed","grifo_con_vecinos_post_razon.RDS")) -->
<!-- ``` -->
