==> R/10_diff-in-diff.Rmd <==
---
title: "Análisis de diferencias en diferencias"
author: "Diego Uriarte"
date: "`r format(Sys.Date())`"
output:
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE, cache = FALSE}
knitr::opts_chunk$set(
    collapse = TRUE,
    comment = "#>",
    error = TRUE,
    message = FALSE,
    warning = FALSE
)
options(tidyverse.quiet = TRUE)

```

```{r  libraries, warning = FALSE, message = FALSE}
library(conflicted)
library(tidyverse)
library(lubridate)
library(stargazer)
library(multiwayvcov)
library(lmtest)
source(here::here("R","funcion04_formato-tablas-reproducibles.R"), encoding = "UTF-8")
'%ni%' <- Negate('%in%')
conflict_prefer("filter", "dplyr")
```

## Cargamos datos y creamos variables


```{r cargar-datos}

data_total <- readRDS(file = here::here("data", "processed", "data-final-regresiones.rds"))

data_total <- data_total %>% 
  mutate(codigo_de_osinergmin = factor(codigo_de_osinergmin),
         mes = (year(fecha)-2017)*12 + month(fecha),
         mes = factor(mes)) %>% 
  mutate(tri_menos_4 = if_else(fecha <= dmy("01-04-2017") & fecha >= dmy("01-02-2017"), 1, 0),
         tri_menos_3 = if_else(fecha <= dmy("01-07-2017") & fecha >= dmy("01-05-2017"), 1, 0),
         tri_menos_2 = if_else(fecha <= dmy("01-10-2017") & fecha >= dmy("01-08-2017"), 1, 0),
         tri_menos_1 = if_else(fecha <= dmy("01-01-2018") & fecha >= dmy("01-11-2017"), 1, 0),
         tri_mas_1 = if_else(fecha <= dmy("01-04-2018") & fecha >= dmy("01-02-2018"), 1, 0),
         tri_mas_2 = if_else(fecha <= dmy("01-07-2018") & fecha >= dmy("01-05-2018"),1,0),
         tri_mas_3 = if_else(fecha <= dmy("01-10-2018") & fecha >= dmy("01-08-2018"),1,0)) 
  
```

## Función para regresiones


Ahora especificamos modelos:

```{r func-reg}
reg_clust <- function(form, 
                      data, 
                      producto, 
                      fecha_lim = c("01-01-2017", "01-10-2018"), 
                      filtrado = FALSE){
  if (filtrado == TRUE) {
    data_filtrada <- balancear_panel(data, prod = producto) %>% 
      filter(fecha >= dmy("01-04-2018") | fecha <= dmy("01-11-2017")) %>% 
      mutate(fecha = factor(fecha)) 
  } else {
    data_filtrada <- balancear_panel(data, prod = producto, fecha_lim) %>% 
      mutate(fecha = factor(fecha))
  }

  modelo <- lm(formula = form, 
                           data = data_filtrada)  
  m1.vcovCL <- cluster.vcov(modelo,data_filtrada$codigo_de_osinergmin)
list(modelo, coeftest(modelo, m1.vcovCL))
}
```

### DiD efecto compra

```{r modelos-compra}

modelo_sin_controles <- precio_de_venta ~ COMPRADA_DID * timing_did

modelo_compradas_ef <- precio_de_venta ~ COMPRADA_DID * timing_did + 
 fecha + distrito

modelo_compradas_controles <- precio_de_venta ~ COMPRADA_DID * timing_did + 
  fecha + 
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon  + distrito 

modelo_tri <- precio_de_venta ~ 
  COMPRADA_DID * tri_menos_1 +
  COMPRADA_DID * tri_mas_1 + 
  COMPRADA_DID * tri_mas_2 + 
  COMPRADA_DID * tri_mas_3 +
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon + distrito +
  fecha  

modelo_tri_test <- precio_de_venta ~ 
  COMPRADA_DID * tri_menos_2 +
  COMPRADA_DID * tri_menos_1 +
  COMPRADA_DID * tri_mas_1 + 
  COMPRADA_DID * tri_mas_2 + 
  COMPRADA_DID * tri_mas_3 +
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon + distrito +
  fecha 

modelo_test_mensual <- precio_de_venta ~ 
  COMPRADA_DID* fecha +
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon + distrito 

modelo_comp <- precio_de_venta ~ 
  COMPRADA_DID* mes +
  tipo_bandera + distrito
  
```


### Diesel efecto compra 


```{r reg-diesel-compra}
diesel_full_sin_controles <- reg_clust(modelo_sin_controles, data_total,
                                       "DIESEL",
                                       fecha_lim = c("01-08-2017","01-10-2018" ))

diesel_full_con_ef <- reg_clust(modelo_compradas_ef, data_total, 
                                "DIESEL",
                                fecha_lim = c("01-08-2017","01-10-2018" ))

diesel_full_controls <- reg_clust(modelo_compradas_controles, data_total, 
                                  "DIESEL",
                                  fecha_lim = c("01-08-2017","01-10-2018" ))

reg_diesel_tri_compra <- reg_clust(modelo_tri, data_total, "DIESEL",
                                   c("01-08-2017","01-10-2018" ))

diesel_tri_test <- reg_clust(modelo_tri_test, data_total, "DIESEL", 
                             c("01-05-2017","01-10-2018" ))

diesel_test_mensual <- reg_clust(modelo_test_mensual, data_total, "DIESEL",
                                 c("01-08-2017","01-10-2018"))

data_total_2 <- data_total %>% 
  filter(fecha <= dmy("01-11-2017") | fecha >= dmy("01-04-2018"))
diesel_test_mensual <- reg_clust(modelo_test_mensual, data_total_2, "DIESEL",
                                 c("01-05-2017","01-10-2018"))
diesel_test_mensual
```


```{r ouput-stargarzer-diesel-compra}
lista_reg <- list(diesel_full_sin_controles, 
                  diesel_full_con_ef, 
                  diesel_full_controls,
                  reg_diesel_tri_compra,
                  diesel_tri_test,
                  diesel_test_mensual)

t <- map(lista_reg, 2)
se <- map(lista_reg, 2) %>% 
  map(~ .[,2])

pe <- map(lista_reg, 2) %>% 
  map(~ .[,4])

stargazer(map(lista_reg, 1),
          type = "html",
          dep.var.labels=str_c("Precio de venta - Diésel (soles/galón)"),
          dep.var.caption = "",
          model.numbers	= T,
          no.space = F,
          # covariate.labels = c("Dum.Comprada", "Dum.FechaCompra", "sc", 
                               # "Dum.Comprada*Dum.FechaCompra", "Constante" ),
          column.labels =  "", 
          se = se,
          p = pe,
          omit = c("tipo_bandera", "fecha", 
                   "codigo_de_osinergmin", "distrito", "^tri_m"),
          # omit.labels = c("Tipo de estación", "Efectos fijos por meses", "Efectos fijos por estación", "EF Distritos"),
          omit.yes.no = c("Sí", "No"),
          omit.stat	= c("f", "ll", 
                        "sigma2", "res.dev", "ser", "aic",
                        "wald", "lr"),
          notes= "Los errores estándares clusterizados por estación se muestran entre paréntesis. ",
          notes.label = "Notas: ",
          single.row = F, 
          out = here::here("doc", "tables", "diesel_did_compra.htm"))
```


### Gasohol 90 compra

```{r reg-gasohol-compra}
g90_full_sin_controles <- reg_clust(modelo_sin_controles, data_total,
                                    "G90",
                                    fecha_lim = c("01-08-2017","01-10-2018" ))

g90_full_con_ef <- reg_clust(modelo_compradas_ef, data_total, 
                             "G90",
                             fecha_lim = c("01-08-2017","01-10-2018" ))

g90_full_controls <- reg_clust(modelo_compradas_controles, data_total, 
                               "G90",
                               fecha_lim = c("01-08-2017","01-10-2018" ))

g90_trimestre_compra <- reg_clust(modelo_tri, data_total, "G90", c("01-08-2017","01-10-2018" ))

g90_tri_test <- reg_clust(modelo_tri_test, data_total, "G90", 
                             c("01-05-2017","01-10-2018" ))
```


```{r output-stargarzer-gasohol-comra}
lista_reg <- list(g90_full_sin_controles, g90_full_con_ef,
                  g90_full_controls, g90_trimestre_compra,
                  g90_tri_test)

t <- map(lista_reg, 2)
se <- map(lista_reg, 2) %>% 
  map(~ .[,2])

pe <- map(lista_reg, 2) %>% 
  map(~ .[,4])

stargazer(map(lista_reg, 1),
          type = "html",
          # dep.var.labels=str_c("Precio de venta - Gasohol 90 (soles/galón)"),
          # dep.var.caption = "",
          model.numbers	= T,
          no.space = T,
          # covariate.labels = c("Dum.Comprada", "Dum.FechaCompra", "sc", 
                               # "Dum.Comprada*Dum.FechaCompra", "Constante" ),
          # column.labels =  "", 
          se = se,
          p = pe,
          omit = c("tipo_bandera", "fecha", 
                   "codigo_de_osinergmin", "distrito", "^tri_m"),
          omit.stat	= c("f", "ll", 
                        "sigma2", "res.dev", "ser", "aic",
                        "wald", "lr"),
          notes= "Los errores estándares clusterizados por estación se muestran entre paréntesis. ",
          notes.label = "Notas: ",
          single.row = F, 
          out = here::here("doc", "tables", "g90_did_compra.htm"))
```


## DiD efecto vecinos

```{r modelos-vecino}

modelo_sin_controles <- precio_de_venta ~ vecino_pecsa_thiessen_did * timing_did

modelo_compradas_ef <- precio_de_venta ~ vecino_pecsa_thiessen_did * timing_did + 
 fecha + distrito

modelo_compradas_controles <- precio_de_venta ~ vecino_pecsa_thiessen_did * timing_did + 
  fecha + 
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon + sc + distrito 

modelo_tri <- precio_de_venta ~ 
  vecino_pecsa_thiessen_did * tri_menos_1 + 
  # vecino_pecsa_thiessen_did * tri_menos_3 + 
  vecino_pecsa_thiessen_did * tri_mas_1 + 
  vecino_pecsa_thiessen_did * tri_mas_2 + 
  vecino_pecsa_thiessen_did * tri_mas_3 +
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon + sc + distrito +
  fecha + 
  vecino_pecsa_thiessen_did * tri_menos_2 
```

### Diesel efecto vecinos 


```{r reg-diesel-vecino}
diesel_full_sin_controles <- reg_clust(modelo_sin_controles, data_total,
                                       "DIESEL",
                                       fecha_lim = c("01-08-2017","01-10-2018" ))

diesel_full_con_ef <- reg_clust(modelo_compradas_ef, data_total, 
                                "DIESEL",
                                fecha_lim = c("01-08-2017","01-10-2018" ))

diesel_full_controls <- reg_clust(modelo_compradas_controles, data_total, 
                                  "DIESEL",
                                  fecha_lim = c("01-08-2017","01-10-2018" ))

reg_diesel_tri_compra <- reg_clust(modelo_tri, data_total, "DIESEL", c("01-08-2017","01-10-2018" ))

```


```{r ouput-stargarzer-diesel-vecinos}
lista_reg <- list(diesel_full_sin_controles, 
                  # diesel_sin3meses_ni_controles,
                  diesel_full_con_ef, 
                  # diesel_full_ef_sc, 
                  diesel_full_controls,
                  reg_diesel_tri_compra)

t <- map(lista_reg, 2)
se <- map(lista_reg, 2) %>% 
  map(~ .[,2])

pe <- map(lista_reg, 2) %>% 
  map(~ .[,4])

stargazer(map(lista_reg, 1),
          type = "html",
          model.numbers	= T,
          no.space = T,
          # covariate.labels = c("Dum.Comprada", "Dum.FechaCompra", "sc", 
                               # "Dum.Comprada*Dum.FechaCompra", "Constante" ),
          column.labels =  "", 
          se = se,
          p = pe,
          omit = c("tipo_bandera", "fecha", 
                   "codigo_de_osinergmin", "distrito", "^tri_m"),
          # omit.labels = c("Tipo de estación", "Efectos fijos por meses", "Efectos fijos por estación", "EF Distritos"),
          omit.yes.no = c("Sí", "No"),
          omit.stat	= c("f", "ll", 
                        "sigma2", "res.dev", "ser", "aic",
                        "wald", "lr"),
          notes= "Los errores estándares clusterizados por estación se muestran entre paréntesis. ",
          notes.label = "Notas: ",
          single.row = F, 
          out = here::here("doc", "tables", "diesel_did_vecino.htm"))
```

### Gasohol 90 vecinos

```{r reg-gasohol-vecino}
g90_full_sin_controles <- reg_clust(modelo_sin_controles, data_total,
                                    "G90",
                                    fecha_lim = c("01-08-2017","01-10-2018" ))

g90_full_con_ef <- reg_clust(modelo_compradas_ef, data_total, 
                             "G90",
                             fecha_lim = c("01-08-2017","01-10-2018" ))

g90_full_controls <- reg_clust(modelo_compradas_controles, data_total, 
                               "G90",
                               fecha_lim = c("01-08-2017","01-10-2018" ))

g90_trimestre_compra <- reg_clust(modelo_tri, data_total, "G90", c("01-08-2017","01-10-2018" ))

```


```{r output-stargarzer-gasohol-vecino}
lista_reg <- list(g90_full_sin_controles, g90_full_con_ef,
                  g90_full_controls, g90_trimestre_compra)

t <- map(lista_reg, 2)
se <- map(lista_reg, 2) %>% 
  map(~ .[,2])

pe <- map(lista_reg, 2) %>% 
  map(~ .[,4])

stargazer(map(lista_reg, 1),
          type = "html",
          # dep.var.labels=str_c("Precio de venta - Gasohol 90 (soles/galón)"),
          # dep.var.caption = "",
          model.numbers	= T,
          no.space = T,
          # covariate.labels = c("Dum.Comprada", "Dum.FechaCompra", "sc", 
                               # "Dum.Comprada*Dum.FechaCompra", "Constante" ),
          # column.labels =  "", 
          se = se,
          p = pe,
          omit = c("tipo_bandera", "fecha", 
                   "codigo_de_osinergmin", "distrito", "^tri_m"),
          omit.stat	= c("f", "ll", 
                        "sigma2", "res.dev", "ser", "aic",
                        "wald", "lr"),
          order = c("^vecino_pecsa_thiessen_did", "^timing_did",
                    "^vecino_pecsa_thiessen_did:timing_did"),
          notes= "Los errores estándares clusterizados por estación se muestran entre paréntesis. ",
          notes.label = "Notas: ",
          single.row = F, 
          out = here::here("doc", "tables", "g90_did_vecino.htm"))
```

## Evaluamos dependencia espacial con tests de Anselin (1996)

```{r test-anselin}
library(spdep)
library(splm)
# Fixed effects model

tests <- list("lml", "lme", "rlml", "rlme")
prods <- list("DIESEL", "G90")

slmtest_tes <- function(prod, df, tests, fecha_lim = c("01-05-2017", "01-10-2018")) {
    data_periodo <- balancear_panel(df, prod, fecha_lim)
    w_prod <- crear_spatial_w(data_periodo, prod)
    slmtest(modelo_compradas_controles, data = data_periodo, listw = w_prod, test=tests,
            model="within")   
}
data_periodo <- balancear_panel(data_total, "G90", c("01-08-2017", "01-10-2018"))

w_prod <- crear_spatial_w(data_periodo, "DIESEL")
```

```{r}
grifos_creados_luego <- data_periodo %>%
  filter(producto == "G90") %>%
  count(codigo_de_osinergmin, sort = T) %>%
  filter(n < max(n)) %>%
  pull(codigo_de_osinergmin)

df_balanceado <- data_periodo %>%
  filter(producto == "G90",
         codigo_de_osinergmin %ni% grifos_creados_luego) 

coords_db5 <- data_periodo %>%
  distinct(codigo_de_osinergmin, .keep_all = T) %>%
  select(codigo_de_osinergmin, lon, lat) %>%
  mutate_at(c("lon", "lat"), as.numeric) %>% 
  mutate_at("codigo_de_osinergmin", as.character)
    
    #creamos la matriz de distancias
grifos_nb <- tri2nb(as.matrix(coords_db5[,2:3]), row.names = coords_db5$codigo_de_osinergmin)
str(grifos_nb)
class(grifos_nb)
sp_grifos <- nb2listw(grifos_nb)
my_matrix <- nb2mat(sp_grifos$neighbours)


```

```{r}


modelo_compradas_ef <- precio_de_venta ~ vecino_pecsa_thiessen_did * timing_did + distrito #+ distrito

modelo_compradas_controles <- precio_de_venta ~ vecino_pecsa_thiessen_did * timing_did + 
  # fecha + 
  tipo_bandera + grifo_mas_cercano + distancia_min + num_grifos_cerc +
  tiene_mecanico + densidad_2017 + ingresos_2012 + 
  num_viajes_millon + sc + distrito 


slmtest(modelo_compradas_controles, data = df_balanceado, listw = sp_grifos, test="lme",
            model="pooling", index = c("codigo_de_osinergmin", "fecha"))   
```

```{r}
library(plm)
x <- plm(modelo_compradas_ef, data = data_periodo, model = "pooling", effect = "twoways",
         index = c("codigo_de_osinergmin", "fecha") )
summary(x)
slmtest(x, listw = sp_grifos, test = "rlme")


sararfemod <- spml(formula = modelo_compradas_ef, 
                   data = data_periodo, 
                   index = c("codigo_de_osinergmin", "fecha"),
                   listw = sp_grifos, 
                   lag = TRUE, spatial.error = "none", model = "pooling",
                   effect = "individual", method = "eigen", na.action = na.fail,
                   quiet = TRUE, zero.policy = NULL, hess = FALSE,
                   tol.solve = 1e-13, control = list(), legacy = FALSE)

summary(sararfemod)
```

```{r}
modelo_compradas_ef <- precio_de_venta ~ COMPRADA_DID * timing_did + 
  distrito

modelo_compradas_controles <- precio_de_venta ~ COMPRADA_DID * timing_did + 
   grifo_mas_cercano  + distancia_min + num_grifos_cerc #+
   tiene_mecanico + densidad_2017 + ingresos_2012 
  # num_viajes_millon + sc + distrito 

t <- spml(formula =  modelo_compradas_ef, 
                   data = data_periodo, 
                   index = c("codigo_de_osinergmin", "fecha"),
                   listw = sp_grifos, 
                   lag = TRUE, spatial.error = "none", model = "pooling",
                   effect = "twoways", method = "eigen", na.action = na.fail,
                   quiet = TRUE, zero.policy = NULL, hess = FALSE,
                   tol.solve = 1e-13, control = list(), legacy = FALSE)

summary(t)
imp1 <- impacts(t, listw = sp_grifos, time = 21, R = 100)
summary(imp1, zstats = TRUE, short = T)

```
