==> R/11_DiD-package.Rmd <==
---
title: "11_DiD-package"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Cargamos librerías

```{r}
library(did)
library(tidyverse)
library(lubridate)
source(here::here("R","funcion05_balancear-panel.R"), encoding = "UTF-8")
'%ni%' <- Negate('%in%')
data(mpdta)
```


## Variables base


```{r cargar-datos}

data_total <- readRDS(file = here::here("data", "processed", "data-final-regresiones.rds"))

distritos_con_pecsa <- data_total %>% 
  distinct(distrito, tipo_bandera) %>% 
  filter(tipo_bandera == "PROPIA PECSA") %>% 
  pull(distrito)


distritos_con_primax <- data_total %>% 
  distinct(distrito, tipo_bandera) %>% 
  filter(tipo_bandera == "PROPIA PRIMAX") %>% 
  pull(distrito)

lista_vecinos <- data_total %>% 
filter(vecino_pecsa_thiessen_did == 1) %>%
  distinct(codigo_de_osinergmin) %>% 
  pull()

lista_no_vecino <- data_total %>% 
filter(vecino_pecsa_thiessen_did == 0) %>%
  distinct(codigo_de_osinergmin) %>% 
  pull()

lista_vecinos_km <- data_total %>% 
filter(vecino_pecsa_dist_did == 1) %>%
  distinct(codigo_de_osinergmin) %>% 
  pull()

```

## Caso base, sin controles y PECSA versus todas las demás


1) Regresión similar a la corrida solo con el paqeute base.
- Plazo recortado de tiempo
- No se define adecuadamente el grupo de control

```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-05-2017", "01-10-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))

out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~1,
              data = data_g90,
              est_method = "reg",
              anticipation = 3,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.3,.3))
summary(out_1)
```

No se notan casi efectos.

2) Lo mismo, cambiando intervalos


```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-09-2017", "01-09-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~1,
              data = data_g90,
              est_method = "reg",
              anticipation = 3,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.3,.3))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

No se detectan efectos claros nuevamente.

3) Intentamos agregando controles

- Con codigo de osinergmin y bandera mejora
```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-09-2017", "01-09-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + bandera,
              data = data_g90,
              est_method = "reg",
              anticipation = 3,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.5,.5))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

4) Con codigo de osinergmin y bandera mejora y distrito, empeora

```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-09-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + bandera + distrito,
              data = data_g90,
              est_method = "reg",
              anticipation = 3,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.5,.5))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

5) Pero en ambos casos le metemos solo dos periodos de anticipación, y mejora 
bastante, en ambos casos ya se parece. Y mejor si le ponemos toda la muestra

```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-09-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + bandera + distrito,
              data = data_g90,
              est_method = "reg",
              anticipation = 2,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.5,.5))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```


6) Añadimos a la especificación anterior, tiene mecánico

```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-09-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + bandera + distrito +
                tiene_mecanico ,
              data = data_g90,
              est_method = "reg",
              anticipation = 2,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.5,.5))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```
No mejora mucho la situación. 

Resumen: La mejor fue la opción 5).

## Probamos para el diesel

Balanceamos el panel:

```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-09-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
```

```{r}
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + distrito + bandera,
              data = data_diesel,
              est_method = "reg"
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```


Cogemos la que mejor funcionó para G90 con Diésel

```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-09-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0))
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = F
              )
ggdid(out_1, ylim = c(-.5,.5))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)

# conditional_did_pretest("precio_de_venta", "mes", "codigo_de_osinergmin", 
#                         "fecha_trat", xformla=~codigo_de_osinergmin + bandera + distrito,
#                         data=data_diesel)
```


# Ahora probemos ajustando mejor el grupo de control

1) Para la gasolina, comparemos los precios de PECSA, versus los precios de los 
distritos donde no está PECSA. Retiramos también PRIMAX


```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-10-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | distrito %ni% distritos_con_pecsa)

out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + bandera,
              data = data_g90,
              est_method = "reg",
              anticipation = 2,
              bstrap = F,
              cband = T
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```
Resultado: Aceptable con anticipacion 2. No corre si controlo por distritos

2) Lo mismo que lo anterior, pero ahora meto todas las estaciones que no son
vecinas

```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-10-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | codigo_de_osinergmin %ni% lista_vecinos)
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_g90,
              est_method = "reg",
              anticipation = 2,
              bstrap = F,
              cband = T
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

Se ve super bien.

3) Si no le pongo anticipación se ve igual de bien:

```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-10-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | codigo_de_osinergmin %ni% lista_vecinos)
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin + bandera,
              data = data_g90,
              est_method = "dr",
              anticipation = 0,
              bstrap = F,
              cband = F,
              )
# dp <- pre_process_did(yname = "precio_de_venta",
#               gname = "fecha_trat",
#               idname = "codigo_de_osinergmin",
#               tname = "mes",
#               xformla = ~codigo_de_osinergmin + distrito,
#               data = data_g90,
#               allow_unbalanced_panel = F)
# 
# results <- compute.att_gt(dp)
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

Por que la diferencia en los meses 6, 9 y 12?


Comparemos los precios entre ambos grupos por meses:

```{r}
plot_precios_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | codigo_de_osinergmin %ni% lista_vecinos_km) %>% 
  mutate(pecsa = if_else(tipo_bandera == "PROPIA PECSA", "PECSA", "NO VECINA")) %>% 
  group_by(pecsa, mes) %>% 
  summarize(precio_prom = mean(precio_de_venta))

plot_precios_g90 %>% 
  ggplot(aes(x = mes, y = precio_prom, color = pecsa)) + 
  geom_point()+
  geom_line()+
  geom_text(aes(label = mes), hjust=-0.002, vjust = -0.03)+
  labs(title= "Efectos en el precio de G90 para estaciones de PECSA")
```

Armemos la info para el HonestDiD

```{r}
library(HonestDiD)
data('LWdata_EventStudy', package = "HonestDiD")
betahat <- out_1$att
sigma <- as.matrix(out_1$V_analytical)
timeVec = c(seq(from = 1, to = 13, by = 1), seq(from = 15, to = 22, by = 1))
referencePeriod = 14
postPeriodIndices = which(timeVec > 14)
prePeriodIndices = which(timeVec < 14)
stdErrors = out_1$se
EventStudy = list(
  betahat = betahat,
  sigma = sigma,
  timeVec = timeVec,
  referencePeriod = referencePeriod,
  prePeriodIndices = prePeriodIndices,
  postPeriodIndices = postPeriodIndices,
  stdErrors = stdErrors)
```

```{r}
#Create l_vec corresponding with 15 years of exposure

l_vec = basisVector(5 , 21-13+1)
# Construct robust confidence intervals for Delta^{SD}(M) for 15 years of exposure
DeltaSD_RobustResults <- createSensitivityResults(
  betahat = EventStudy$betahat,
  sigma = EventStudy$sigma,
  numPrePeriods = 12,
  numPostPeriods = 21-13+1,
  l_vec = l_vec,
  Mvec = seq(from = 0, to = 0.025, by =0.005) 
)


OriginalResults <- constructOriginalCS(
  betahat = EventStudy$betahat,
  sigma = EventStudy$sigma,
  numPrePeriods = 12,
  numPostPeriods = 21-13+1,
  l_vec = l_vec
)

DeltaSD_SensitivityPlot <- createSensitivityPlot(
  robustResults = DeltaSD_RobustResults,
  originalResults = OriginalResults
)
DeltaSD_SensitivityPlot

head(DeltaSD_RobustResults)
```

DIESEL 

4) Lo mismo, para el diesel
Ojo, si intento controlar por bandera, se incrementa los IC

```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | codigo_de_osinergmin %ni% lista_vecinos)
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 2,
              bstrap = F,
              cband = T
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)

```

5) Nuavemente, con cero se obtienen resultados tambien se bien

```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | codigo_de_osinergmin %ni% lista_vecinos)
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = T
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)

```
5) Similar, pero con las estaciones en distritos donde no está pecsa

```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PECSA", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PRIMAX",
        tipo_bandera == "PROPIA PECSA" | distrito %ni% distritos_con_pecsa)
        #fecha >= dmy("01/03/2018") | fecha <= dmy("01-09-2017"))
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = F
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)

```

No cambia mucho.

Conclusión, poner anticipación 0 requiere menos malares y se bien igual de bien.

========================================================


## Para Primax también hay efectos interesantes

1) Efecto en el precio para estaciones de Primax, descartando las de pecsa, y con control
aquellas en distritos donde no está Primax
```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-10-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PRIMAX", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PECSA",
        tipo_bandera == "PROPIA PRIMAX" | distrito %ni% distritos_con_primax)
        #fecha >= dmy("01/03/2018") | fecha <= dmy("01-09-2017"))
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_g90,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = F
              )
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

2) Efecto en el precio de diesel para estaciones de Primax, descartando las de pecsa, y con control
```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         fecha_trat = if_else(tipo_bandera == "PROPIA PRIMAX", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PECSA",
        tipo_bandera == "PROPIA PRIMAX" | distrito %ni% distritos_con_primax)
        #fecha >= dmy("01/03/2018") | fecha <= dmy("01-09-2017"))
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 0,
              bstrap = T,
              cband = T
              )
ggdid(out_1)
summary(out_1)
```

Para los precios mismos de Primax, se nota claramente el incremento

# Vecinas 
5) Vecinas de Pecsa en Gasolina
```{r}
data_g90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-10-2018"))
data_g90 <- data_g90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         vecina = if_else(codigo_de_osinergmin %in% lista_vecinos, 
                          "vecina", 
                          "no vecina"),
         fecha_trat = if_else(vecina == "vecina", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PECSA",
         tipo_bandera != "PROPIA PRIMAX")
        #tipo_bandera == "PROPIA PRIMAX" | distrito %ni% distritos_con_primax)
        #fecha >= dmy("01/03/2018") | fecha <= dmy("01-09-2017"))
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_g90,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = T)
ggdid(out_1)
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

Efectos positivos hacia el final, al inicio no. Pero se ven

6) Vecinas de Pecsa, efecto en precio de Diesel

```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         vecina = if_else(codigo_de_osinergmin %in% lista_vecinos, 
                          "vecina", 
                          "no vecina"),
         fecha_trat = if_else(vecina == "vecina", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PECSA",
         tipo_bandera != "PROPIA PRIMAX")
        #tipo_bandera == "PROPIA PRIMAX" | distrito %ni% distritos_con_primax)
        #fecha >= dmy("01/03/2018") | fecha <= dmy("01-09-2017"))
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = T)
ggdid(out_1, ylim = c(-.2,.2))
summary(out_1)
group_effects <- aggte(out_1, type = "group")
summary(group_effects)
```

No hay efecto global, ni hacia el final

7) Comparamos las estaciones donde no estaba Primax, contra las que estaba
```{r}
data_diesel <- balancear_panel(data_total, prod = "DIESEL", c("01-01-2017", "01-10-2018"))
data_diesel <- data_diesel %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         vecina = if_else(distrito %in% distritos_con_primax, 
                          "vecina", 
                          "no vecina"),
         fecha_trat = if_else(vecina == "vecina", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PECSA",
         tipo_bandera != "PROPIA PRIMAX")
           
out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_diesel,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = T
              )
ggdid(out_1)
summary(out_1)
```

8) Efecto en el precio de G90 para vecinas de Primax, descartando las de pecsa, y con control
```{r}
data_G90 <- balancear_panel(data_total, prod = "G90", c("01-01-2017", "01-10-2018"))
data_G90 <- data_G90 %>% 
  mutate(mes = (year(fecha)-2017)*12 + month(fecha),
         vecina = if_else(distrito %in% distritos_con_primax, 
                          "vecina", 
                          "no vecina"),
         fecha_trat = if_else(vecina == "vecina", 14, 0)) %>% 
  filter(tipo_bandera != "PROPIA PECSA",
        tipo_bandera != "PROPIA PRIMAX" )

out_1 <- att_gt(yname = "precio_de_venta",
              gname = "fecha_trat",
              idname = "codigo_de_osinergmin",
              tname = "mes",
              xformla = ~codigo_de_osinergmin,
              data = data_G90,
              est_method = "reg",
              anticipation = 0,
              bstrap = F,
              cband = T
              )
ggdid(out_1)
summary(out_1)
```

# HonestDiD

```{r}
data('LWdata_EventStudy', package = "HonestDiD")

```
